requirements.txt
# Core
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.0.5
openai>=1.30.0
python-dotenv>=1.0.0

# SQL
sqlalchemy>=2.0
# Databricks dialect/driver
databricks-sql-connector>=3.0

#neo4j
sqlalchemy-databricks

